pyproject.toml
[project]
name = "gen-eval"
version = "0.0.1"
description = "Project GenEval: A Unified Evaluation Framework for Generative AI Applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0.0",
    "datasets>=4.0.0",
    "deepeval>=3.3.2",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "ragas>=0.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "langchain-community>=0.1.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "pytest-cov>=6.2.1",
    "regex>=2025.2.10",
    "safety>=3.2.11",
]

[project.scripts]
geneval = "geneval.cli:cli"

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "pip-audit>=2.6.0",
    "bandit>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["geneval"]

[tool.pytest.ini_options]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]

[tool.black]
line-length = 150
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 150
known_first_party = ["geneval"]

[tool.ruff]
line-length = 150
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B006",  # do not use mutable data structures for argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["geneval"]
